+++ src/modules/ImprovedGroupView.cpp
@@ -91,7 +91,13 @@ void IGVSetGroupIDLayer::onRemoveFromGroup2(CCObject* obj) {
 void IGVSetGroupIDLayer::regenerateGroupView() {
     auto fields = m_fields.self();
 
-    if (fields->m_scrollLayer) fields->m_scrollLayer->removeFromParent();
+    if (fields->m_scrollLayer) {
+        fields->m_scrollLayer->removeFromParent();
+    }
+
+    if (fields->m_scrollBar) {
+        fields->m_scrollBar->removeFromParent();
+    }
 
     std::vector<ImprovedGroupView::GroupData> groupData;
     std::map<int, int> allGroups;
@@ -236,9 +242,27 @@ void IGVSetGroupIDLayer::regenerateGroupView() {
 
     m_mainLayer->addChild(fields->m_scrollLayer);
 
-    if (menuContainer->getScaledContentHeight() <= 67) {
+    fields->m_scrollBar = AdvancedScrollBar::create(fields->m_scrollLayer, ScrollOrientation::VERTICAL);
+    auto style = RoundedScrollStyle();
+    style.m_track = [] () {
+        auto track = RoundedScrollTrack::create();
+        track->setClickColor({0, 0, 0, 0});
+        track->setBackgroundColor({0, 0, 0, 0});
+        return track;
+    };
+
+    fields->m_scrollBar->setStyle(style);
+    fields->m_scrollBar->setContentSize({10, fields->m_scrollLayer->getContentHeight() - 10});
+    fields->m_scrollBar->setPosition({fields->m_scrollLayer->getPositionX() + fields->m_scrollLayer->getContentWidth() / 2 - fields->m_scrollBar->getContentWidth() / 2 + 1, fields->m_scrollLayer->getPositionY()});
+    fields->m_scrollBar->setZOrder(100);
+    fields->m_scrollBar->setTouchPriority(fields->m_scrollLayer->getTouchPriority() - 10);
+
+    m_mainLayer->addChild(fields->m_scrollBar);
+
+    if (groupsMenu->getScaledContentHeight() <= 67) {
         fields->m_scrollLayer->setVerticalScroll(false);
         fields->m_scrollLayer->setHorizontalScrollWheel(false);
+        fields->m_scrollBar->setVisible(false);
     }
 
     if (fields->m_groupCountLabel) fields->m_groupCountLabel->removeFromParent();
@@ -280,4 +304,412 @@ ImprovedGroupView::GroupData IGVSetGroupIDLayer::parseObjGroups(GameObject* obj)
     }
 
     return ImprovedGroupView::GroupData{groups, parents, obj};
+}
+
+void IGVSetupSpawnPopup::addRemap(int oldID, int newID) {
+    SetupSpawnPopup::addRemap(oldID, newID);
+    m_fields->m_needsUpdate = true;
+}
+
+void IGVSetupSpawnPopup::onDeleteRemap(cocos2d::CCObject* sender) {
+    SetupSpawnPopup::onDeleteRemap(sender);
+    m_fields->m_needsUpdate = true;
+}
+
+void IGVSetupSpawnPopup::queueUpdateButtons() {
+    SetupSpawnPopup::queueUpdateButtons();
+    m_fields->m_needsUpdate = true;
+}
+
+bool IGVSetupSpawnPopup::init(EffectGameObject* object, cocos2d::CCArray* objects) {
+    if (!SetupSpawnPopup::init(object, objects)) return false;
+
+    if (tinker::utils::getMod<"spaghettdev.named-editor-groups">()) {
+        schedule(schedule_selector(IGVSetupSpawnPopup::fixNamedEditorGroups));
+    }
+
+    return true;
+}
+
+void IGVSetupSpawnPopup::fixNamedEditorGroups(float dt) {
+    auto fields = m_fields.self();
+    if (fields->m_lastPage == m_page) return;
+    fields->m_lastPage = m_page;
+
+    if (m_page == 1) {
+        if (tinker::utils::getMod<"spaghettdev.named-editor-groups">()) {
+            runAction(CallFuncExt::create([this] {
+                queueUpdateButtons();
+            }));
+        } 
+    }
+}
+
+void IGVSetupSpawnPopup::updateRemapButtons(float dt) {
+    auto fields = m_fields.self();
+
+    if (!fields->m_needsUpdate) return;
+    fields->m_needsUpdate = false;
+
+    m_isBusy = false;
+
+    for (auto btn : CCArrayExt<CCNode, false>(m_remapButtons)) {
+        btn->removeFromParent();
+    }
+
+    auto namedEditorGroupsMenu = m_buttonMenu->getChildByID("spaghettdev.named-editor-groups/remaps-list-menu");
+    if (namedEditorGroupsMenu) {
+        namedEditorGroupsMenu->setVisible(false);
+    }
+
+    m_remapButtons->removeAllObjects();
+    m_remapGroups.clear();
+
+    float currentScroll = 0;
+
+    if (fields->m_scrollLayer) {
+        currentScroll = fields->m_scrollLayer->getScrollPoint().y;
+        fields->m_scrollLayer->removeFromParent();
+        removeObjectFromPage(fields->m_scrollLayer, 1);
+    }
+
+    if (fields->m_scrollBar) {
+        fields->m_scrollBar->removeFromParent();
+        removeObjectFromPage(fields->m_scrollBar, 1);
+    }
+
+    if (fields->m_groupCountLabel) {
+        fields->m_groupCountLabel->removeFromParent();
+        removeObjectFromPage(fields->m_groupCountLabel, 1);
+    }
+
+    auto winSize = CCDirector::get()->getWinSize();
+
+    auto groupsMenu = CCMenu::create();
+
+    auto objects = m_gameObject ? CCArray::createWithObject(m_gameObject) : m_gameObjects;
+
+    struct RemapEntry {
+        int groupID;
+        int remapID;
+
+        bool operator<(const RemapEntry& rhs) const {
+            return std::tie(groupID, remapID) < std::tie(rhs.groupID, rhs.remapID);
+        }
+    };
+
+    std::map<RemapEntry, int> remapCounts;
+
+    for (auto obj : CCArrayExt<GameObject*>(objects)) {
+        for (const auto& pair : static_cast<SpawnTriggerGameObject*>(obj)->m_remapObjects) {
+            RemapEntry entry {pair.m_groupID, pair.m_chance};
+            remapCounts[entry]++;
+        }
+    }
+
+    int index = 0;
+
+    for (const auto& [entry, count] : remapCounts) {
+        auto sprite = entry.groupID == m_remapOriginalID && entry.remapID == m_remapNewID ? "GJ_button_03.png" : (count == objects->count() ? "GJ_button_04.png" : "GJ_button_05.png");
+
+        auto buttonSprite = ButtonSprite::create(
+            fmt::format("{}\n{}", entry.groupID, entry.remapID).c_str(),
+            remapCounts.size() > 10 ? 15 : 40,
+            0,
+            0.35f,
+            true,
+            "bigFont.fnt",
+            sprite,
+            30.f
+        );
+
+        auto button = CCMenuItemSpriteExtra::create(
+            buttonSprite,
+            this,
+            menu_selector(SetupSpawnPopup::onSelectRemap)
+        );
+
+        button->setTag(index);
+
+        groupsMenu->addChild(button);
+
+        m_remapButtons->addObject(button);
+
+        m_remapGroups.emplace_back(entry.groupID);
+        m_remapGroups.emplace_back(entry.remapID);
+
+        index++;
+    }
+
+    auto layout = RowLayout::create();
+    layout->setGap(10);
+    layout->setAutoScale(false);
+    layout->setGrowCrossAxis(true);
+    layout->setCrossAxisOverflow(true);
+    layout->setPadding({20, 10, 20, 10});
+    
+    if (ImprovedGroupView::getSetting<bool, "left-align">()) {
+        layout->setAxisAlignment(AxisAlignment::Start);
+    }
+
+    groupsMenu->setLayout(layout);
+
+    groupsMenu->setContentSize({360, 90});
+    groupsMenu->setPosition({360/2.f, 0});
+    groupsMenu->setAnchorPoint({0.5, 0});
+    groupsMenu->updateLayout();
+    groupsMenu->setID("groups-menu"_spr);
+
+    if (groupsMenu->getScaledContentHeight() > 90) {
+        layout->setPadding({20, 5, 20, 5});
+        groupsMenu->updateLayout();
+    }
+
+    fields->m_scrollLayer = AdvancedScrollLayer::create({360, 90});
+
+    fields->m_scrollLayer->setContentSize({360, 90});
+    fields->m_scrollLayer->setPosition({winSize.width/2, winSize.height/2});
+    fields->m_scrollLayer->ignoreAnchorPointForPosition(false);
+    fields->m_scrollLayer->addChild(groupsMenu);
+    fields->m_scrollLayer->setID("groups-list-scroll-layer"_spr);
+    fields->m_scrollLayer->setTouchPriority(-504);
+    fields->m_scrollLayer->setScrollDelta(1.5f);
+    fields->m_scrollLayer->getContentLayer()->setLayout(SimpleRowLayout::create()->setCrossAxisScaling(AxisScaling::Grow)->setCrossAxisAlignment(CrossAxisAlignment::Start));
+
+    m_mainLayer->addChild(fields->m_scrollLayer);
+
+    fields->m_scrollLayer->getContentLayer()->updateLayout();
+    fields->m_scrollLayer->setScrollY(currentScroll, false);
+
+    fields->m_groupCountLabel = CCLabelBMFont::create(fmt::format("Remap Groups: {}", remapCounts.size()).c_str(), "chatFont.fnt");
+
+    fields->m_groupCountLabel->setID("group-count-label"_spr);
+    fields->m_groupCountLabel->setAnchorPoint({0, 1.f});
+    fields->m_groupCountLabel->setColor({0, 0, 0});
+    fields->m_groupCountLabel->setOpacity(200);
+    fields->m_groupCountLabel->setScale(0.5f);
+    fields->m_groupCountLabel->setPosition({fields->m_scrollLayer->getPositionX() - fields->m_scrollLayer->getContentWidth()/2, fields->m_scrollLayer->getPositionY() - fields->m_scrollLayer->getContentHeight()/2 - 2});
+
+    m_mainLayer->addChild(fields->m_groupCountLabel);
+
+    fields->m_scrollBar = AdvancedScrollBar::create(fields->m_scrollLayer, ScrollOrientation::VERTICAL);
+    auto style = RoundedScrollStyle();
+    style.m_track = [] () {
+        auto track = RoundedScrollTrack::create();
+        track->setClickColor({0, 0, 0, 0});
+        track->setBackgroundColor({0, 0, 0, 0});
+        return track;
+    };
+
+    fields->m_scrollBar->setStyle(style);
+    fields->m_scrollBar->setContentSize({10, fields->m_scrollLayer->getContentHeight() - 10});
+    fields->m_scrollBar->setPosition({fields->m_scrollLayer->getPositionX() + fields->m_scrollLayer->getContentWidth() / 2 - fields->m_scrollBar->getContentWidth() / 2 + 1, fields->m_scrollLayer->getPositionY()});
+    fields->m_scrollBar->setZOrder(100);
+    fields->m_scrollBar->setTouchPriority(fields->m_scrollLayer->getTouchPriority() - 10);
+
+    m_mainLayer->addChild(fields->m_scrollBar);
+
+    if (groupsMenu->getScaledContentHeight() <= 90) {
+        fields->m_scrollLayer->setVerticalScroll(false);
+        fields->m_scrollLayer->setHorizontalScrollWheel(false);
+        fields->m_scrollBar->setVisible(false);
+    }
+
+    addObjectToPage(fields->m_scrollBar, 1);
+    addObjectToPage(fields->m_groupCountLabel, 1);
+    addObjectToPage(fields->m_scrollLayer, 1);
+}
+
+void IGVSetupRandAdvTriggerPopup::updateGroupIDButtons() {
+    auto fields = m_fields.self();
+
+    for (auto btn : CCArrayExt<CCNode, false>(m_groupButtons)) {
+        btn->removeFromParent();
+    }
+
+    auto namedEditorGroupsMenu = m_buttonMenu->getChildByID("spaghettdev.named-editor-groups/groups-list-menu");
+    if (namedEditorGroupsMenu) {
+        namedEditorGroupsMenu->setVisible(false);
+    }
+
+    m_groupButtons->removeAllObjects();
+
+    float currentScroll = 0;
+
+    if (fields->m_scrollLayer) {
+        currentScroll = fields->m_scrollLayer->getScrollPoint().y;
+        fields->m_scrollLayer->removeFromParent();
+    }
+
+    if (fields->m_scrollBar) {
+        fields->m_scrollBar->removeFromParent();
+    }
+
+    if (fields->m_groupCountLabel) {
+        fields->m_groupCountLabel->removeFromParent();
+    }
+
+    std::vector<ChanceObject> chanceObjects;
+
+    auto obj = static_cast<RandTriggerGameObject*>(m_gameObject);
+    auto totalChance = obj->getTotalChance();
+
+    for (auto chance : obj->m_chanceObjects) {
+        chance.m_unk00c = totalChance;
+        chanceObjects.push_back(chance);
+    }
+    
+    std::sort(chanceObjects.begin(), chanceObjects.end(), [] (const ChanceObject& left, const ChanceObject& right) {
+        return left.m_groupID < right.m_groupID;
+    });
+
+    auto groupsMenu = CCMenu::create();
+
+    for (auto& obj : chanceObjects) {
+
+        std::string sprite = "GJ_button_04.png";
+        if (!m_gameObject && obj.m_chance < 1) {
+            sprite = "GJ_button_05.png";
+        }
+
+        std::string label;
+
+        if (obj.m_chance < 1) {
+            label = fmt::format("{} - ?\n?%", obj.m_groupID);
+        }
+        else {
+            int percent = static_cast<float>(obj.m_chance) / obj.m_unk00c * 100.f;
+            label = fmt::format("{} - {}\n{}%", obj.m_groupID, obj.m_chance, percent);
+        }
+
+        auto size = CCSize{40.f, 30.f};
+        bool hasName = false;
+        if (tinker::utils::getMod<"spaghettdev.named-editor-groups">()) {
+            size.height = 35.f;
+
+            std::string name = NIDManager::getNameForID(NID::GROUP, obj.m_groupID).unwrapOrDefault();
+            if (!name.empty()) {
+                label += fmt::format("\n{}", name);
+                hasName = true;
+                size.width = 60.f;
+            }
+        }
+        
+        auto button = ButtonSprite::create(
+            label.c_str(),
+            size.width,
+            0,
+            0.35f,
+            true,
+            "bigFont.fnt",
+            sprite.c_str(),
+            size.height
+        );
+
+        if (obj.m_chance < 1) {
+            button->m_label->setColor({255, 150, 0});
+        }
+
+        if (hasName) {
+            if (button->m_label->getScale() >= (0.35f * 0.8f)) {
+                button->m_label->setScale(button->m_label->getScale() * 0.8f);
+            }
+            button->m_label->setPositionX(9 + button->m_label->getScaledContentWidth() / 2);
+        }
+
+        auto menuItem = CCMenuItemSpriteExtra::create(
+            button,
+            this,
+            menu_selector(SetupRandAdvTriggerPopup::onRemoveFromGroup)
+        );
+
+        menuItem->setTag(obj.m_groupID);
+
+        groupsMenu->addChild(menuItem);
+        m_groupButtons->addObject(menuItem);
+    }
+
+    auto layout = RowLayout::create();
+    layout->setGap(10);
+    layout->setAutoScale(false);
+    layout->setGrowCrossAxis(true);
+    layout->setCrossAxisOverflow(true);
+    layout->setPadding({15, 10, 15, 10});
+
+    if (ImprovedGroupView::getSetting<bool, "left-align">()) {
+        layout->setAxisAlignment(AxisAlignment::Start);
+    }
+
+    groupsMenu->setLayout(layout);
+
+    groupsMenu->setContentSize({350, 90});
+    groupsMenu->setPosition({350/2.f, 0});
+    groupsMenu->setAnchorPoint({0.5, 0});
+    groupsMenu->updateLayout();
+    groupsMenu->setID("groups-menu"_spr);
+
+    if (groupsMenu->getScaledContentHeight() > 90) {
+        layout->setPadding({15, 5, 15, 5});
+        groupsMenu->updateLayout();
+    }
+
+    auto winSize = CCDirector::get()->getWinSize();
+
+    fields->m_scrollLayer = AdvancedScrollLayer::create({350, 90});
+
+    fields->m_scrollLayer->setContentSize({350, 90});
+    fields->m_scrollLayer->setPosition({winSize.width/2, winSize.height/2 - 10});
+    fields->m_scrollLayer->ignoreAnchorPointForPosition(false);
+    fields->m_scrollLayer->addChild(groupsMenu);
+    fields->m_scrollLayer->setID("groups-list-scroll-layer"_spr);
+    fields->m_scrollLayer->setTouchPriority(-504);
+    fields->m_scrollLayer->setScrollDelta(1.5f);
+    fields->m_scrollLayer->getContentLayer()->setLayout(SimpleRowLayout::create()->setCrossAxisScaling(AxisScaling::Grow)->setCrossAxisAlignment(CrossAxisAlignment::Start));
+
+    m_mainLayer->addChild(fields->m_scrollLayer);
+
+    fields->m_scrollLayer->getContentLayer()->updateLayout();
+    fields->m_scrollLayer->setScrollY(currentScroll, false);
+
+    fields->m_scrollBar = AdvancedScrollBar::create(fields->m_scrollLayer, ScrollOrientation::VERTICAL);
+    auto style = RoundedScrollStyle();
+    style.m_track = [] () {
+        auto track = RoundedScrollTrack::create();
+        track->setClickColor({0, 0, 0, 0});
+        track->setBackgroundColor({0, 0, 0, 0});
+        return track;
+    };
+
+    fields->m_scrollBar->setStyle(style);
+    fields->m_scrollBar->setContentSize({10, fields->m_scrollLayer->getContentHeight() - 10});
+    fields->m_scrollBar->setPosition({fields->m_scrollLayer->getPositionX() + fields->m_scrollLayer->getContentWidth() / 2 - fields->m_scrollBar->getContentWidth() / 2 + 1, fields->m_scrollLayer->getPositionY()});
+    fields->m_scrollBar->setZOrder(100);
+    fields->m_scrollBar->setTouchPriority(fields->m_scrollLayer->getTouchPriority() - 10);
+
+    m_mainLayer->addChild(fields->m_scrollBar);
+
+    fields->m_groupCountLabel = CCLabelBMFont::create(fmt::format("Chance Groups: {}", chanceObjects.size()).c_str(), "chatFont.fnt");
+
+    fields->m_groupCountLabel->setID("group-count-label"_spr);
+    fields->m_groupCountLabel->setAnchorPoint({0, 1.f});
+    fields->m_groupCountLabel->setColor({0, 0, 0});
+    fields->m_groupCountLabel->setOpacity(200);
+    fields->m_groupCountLabel->setScale(0.5f);
+    fields->m_groupCountLabel->setPosition({fields->m_scrollLayer->getPositionX() - fields->m_scrollLayer->getContentWidth()/2, fields->m_scrollLayer->getPositionY() - fields->m_scrollLayer->getContentHeight()/2 - 2});
+
+    m_mainLayer->addChild(fields->m_groupCountLabel);
+
+    if (groupsMenu->getScaledContentHeight() <= 90) {
+        fields->m_scrollLayer->setVerticalScroll(false);
+        fields->m_scrollLayer->setHorizontalScrollWheel(false);
+        fields->m_scrollBar->setVisible(false);
+    }
+
+    if (tinker::utils::getMod<"spaghettdev.named-editor-groups">()) {
+        runAction(CallFuncExt::create([this] {
+            for (auto btn : CCArrayExt<CCNode, false>(m_groupButtons)) {
+                btn->setVisible(true);
+            }
+        }));
+    }
 }