+++ src/modules/ImprovedGroupView.hpp
@@ -2,6 +2,8 @@
 
 #include "module/Module.hpp"
 #include <Geode/modify/SetGroupIDLayer.hpp>
+#include <Geode/modify/SetupSpawnPopup.hpp>
+#include <Geode/modify/SetupRandAdvTriggerPopup.hpp>
 #include <Geode/modify/EditorUI.hpp>
 #include <alphalaneous.alphas-ui-pack/include/API.hpp>
 
@@ -23,6 +25,7 @@ class $modify(IGVSetGroupIDLayer, SetGroupIDLayer) {
 
     struct Fields {
         AdvancedScrollLayer* m_scrollLayer;
+        AdvancedScrollBar* m_scrollBar;
         CCLabelBMFont* m_groupCountLabel;
 
         int m_lastRemoved = 0;
@@ -34,11 +37,51 @@ class $modify(IGVSetGroupIDLayer, SetGroupIDLayer) {
         (void) self.setHookPriorityAfterPost("SetGroupIDLayer::init", "spaghettdev.named-editor-groups");
     }
     
-    void checkNamedIDs(float dt);
     bool init(GameObject* obj, cocos2d::CCArray* objs);
+    void checkNamedIDs(float dt);
     void onRemoveFromGroup2(CCObject* obj);
     void onAddGroup2(CCObject* obj);
     void onAddGroupParent2(CCObject* obj);
     void regenerateGroupView();
     ImprovedGroupView::GroupData parseObjGroups(GameObject* obj);
+};
+
+class $modify(IGVSetupSpawnPopup, SetupSpawnPopup) {
+    $registerEditorHooks(ImprovedGroupView, true)
+
+    struct Fields {
+        bool m_needsUpdate = true;
+        int m_lastPage = 0;
+        AdvancedScrollLayer* m_scrollLayer;
+        AdvancedScrollBar* m_scrollBar;
+        CCLabelBMFont* m_groupCountLabel;
+    };
+
+    static void _onModify(auto& self) {
+        (void) self.setHookPriorityAfterPost("SetupSpawnPopup::updateRemapButtons", "spaghettdev.named-editor-groups");
+    }
+
+    bool init(EffectGameObject* object, cocos2d::CCArray* objects);
+    void fixNamedEditorGroups(float dt);
+    void addRemap(int oldID, int newID);
+    void onDeleteRemap(cocos2d::CCObject* sender);
+    void queueUpdateButtons();
+    void updateRemapButtons(float dt);
+
+};
+
+class $modify(IGVSetupRandAdvTriggerPopup, SetupRandAdvTriggerPopup) {
+    $registerEditorHooks(ImprovedGroupView, true)
+
+    static void _onModify(auto& self) {
+        (void) self.setHookPriorityAfterPost("SetupRandAdvTriggerPopup::updateGroupIDButtons", "spaghettdev.named-editor-groups");
+    }
+
+    struct Fields {
+        AdvancedScrollLayer* m_scrollLayer;
+        AdvancedScrollBar* m_scrollBar;
+        CCLabelBMFont* m_groupCountLabel;
+    };
+
+    void updateGroupIDButtons();
 };